+++ apple_watch_3/apple_watch_3_test.go
@@ -5,7 +5,8 @@ import (
 	"github.com/golang/mock/gomock"
 	"github.com/gorilla/mux"
 	"github.com/kine-dmd/api/kinesisqueue"
-	"github.com/kine-dmd/api/mocks"
+	"github.com/kine-dmd/api/mocks/mock_kinesis_queue"
+	"github.com/kine-dmd/api/mocks/mock_watch_pos_db"
 	"github.com/kine-dmd/api/watch_position_db"
 	"net/http"
 	"net/http/httptest"
@@ -112,11 +113,11 @@ func TestValidUUID(t *testing.T) {
 	mockCtrl.Finish()
 }
 
-func makeMockQueueAndDB(t *testing.T) (*gomock.Controller, *mocks.MockKinesisQueueInterface, *mocks.MockWatchPositionDB) {
+func makeMockQueueAndDB(t *testing.T) (*gomock.Controller, *mock_kinesisqueue.MockKinesisQueueInterface, *mock_watch_position_db.MockWatchPositionDB) {
 	// Make a mock for the kinesis queue
 	mockCtrl := gomock.NewController(t)
-	mockQueue := mocks.NewMockKinesisQueueInterface(mockCtrl)
-	mockDB := mocks.NewMockWatchPositionDB(mockCtrl)
+	mockQueue := mock_kinesisqueue.NewMockKinesisQueueInterface(mockCtrl)
+	mockDB := mock_watch_position_db.NewMockWatchPositionDB(mockCtrl)
 	return mockCtrl, mockQueue, mockDB
 }
